Add proxy instructions for non-API routes
In development mode, the frontend runs on a separate process, and proxies requests to the backend.

By default, that only happens for paths under /api, under the assumption that the app is a single-page application. When adding backend routes outside that path, it's important to remember that we need to add proxy exceptions too, or the FE will take over and render the react app in development.

This PR adds a comment to the index file to help remind folks about the proxy rules.